feat: add UI for reading and navigating questions
